Change test infrastructure
  some of fixtures/methods is common for all web tests,
    hence the web directory was created in usmqe_tests
  it looks better when the same directory structure is for api tests too,
    hence the api directory was created
